Add spec doc for audit instance queue address
Replaces the initial ServiceControlQueueAddressTests with an executable
specification organized as feature > rule > examples, mirroring the new
docs/specs/audit-instance-servicecontrol-queue-address.md.

The spec captures the four rules for bug #4753 (SCMU does not set
ServiceControl.Audit/ServiceControlQueueAddress when only adding an audit
instance): the instance installed in the same session always wins,
auto-detect when exactly one error instance exists, require an explicit
choice when several exist, and block Save when none exist.

These are outer-loop TDD tests and reference members that do not exist
yet: GetInstalledErrorInstanceNames, ServiceControlQueueAddress,
ServiceControlQueueAddressOptions, ShowServiceControlQueueAddressSelection.
